Python code
I've been trying to make gettaddress() function to get all of the tabs URLs but for now, I only managed to get the only URL of the active tab. If somebody has any idea how to do  it in python please let me know because that would make this a lot more user friendly.